Update opportunity details spec to include tab nav
Includes adding 'spec' to the end of the file name to be in-line with other spec files.
